+++ Controllers/StockController.cs
@@ -27,6 +27,10 @@ namespace api.Controllers
         // return task will be called at the end of the function
         public async Task<IActionResult> GetAll()
         {
+            // modelstate
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stocks = await _stockRepo.GetAllAsync();
             // this is .nets version of map ~ immutable list/array
 
@@ -35,9 +39,13 @@ namespace api.Controllers
             return Ok(stocks);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            // modelstate
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if (stock == null)
@@ -54,16 +62,26 @@ namespace api.Controllers
         // create the request portion of the DTO
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            // modelstate
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = stockDto.ToStockFromCreateDto();
+
             await _stockRepo.CreateAsync(stockModel);
+
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
         // lets make a put
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
+            // modelstate
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
             if(stockModel == null)
@@ -78,10 +96,14 @@ namespace api.Controllers
 
         // DELETE
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
 
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            // modelstate
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel =  await _stockRepo.DeleteAsync(id);
 
             if(stockModel == null)
